+++ extend-search-block.php
@@ -3,10 +3,10 @@
  * Plugin Name:  Extend Search Block
  * Plugin URI:   https://www.webmandesign.eu/portfolio/extend-search-block-wordpress-plugin/
  * Description:  Extending WordPress Search block with search results modifier fields.
- * Version:      1.0.0
+ * Version:      1.0.1
  * Author:       WebMan Design, Oliver Juhas
  * Author URI:   https://www.webmandesign.eu/
- * License:      GNU General Public License v3
+ * License:      GPL-3.0-or-later
  * License URI:  http://www.gnu.org/licenses/gpl-3.0.txt
  * Text Domain:  extend-search-block
  * Domain Path:  /languages
@@ -29,7 +29,7 @@
 defined( 'ABSPATH' ) || exit;
 
 // Constants.
-define( 'EXTEND_SEARCH_BLOCK_VERSION', '1.0.0' );
+define( 'EXTEND_SEARCH_BLOCK_VERSION', '1.0.1' );
 define( 'EXTEND_SEARCH_BLOCK_FILE', __FILE__ );
 define( 'EXTEND_SEARCH_BLOCK_PATH', plugin_dir_path( EXTEND_SEARCH_BLOCK_FILE ) ); // Trailing slashed.
 define( 'EXTEND_SEARCH_BLOCK_URL', plugin_dir_url( EXTEND_SEARCH_BLOCK_FILE ) ); // Trailing slashed.