+++ changelog.md
@@ -1,5 +1,17 @@
 # Extend Search Block changelog
 
+## 1.0.1, 20240508
+
+### Updated
+- Fixing taxonomy dropdown rendering
+
+### File updates
+	changelog.md
+	extend-search-block.php
+	readme.txt
+	includes\Block.php
+
+
 ## 1.0.0, 20230318
 
 - Initial release.